Remove erroneously included installers.
An error on my part, I clicked the download links in the blog posts without
checking that the links were different. Turns out that all of the blog post
links were the same, so old versions weren't preserved.

This means that the following versions don't have an installer: v0.3, v0.5,
v0.7, v0.9, v0.91, and v0.95.

Whoops.