Include the correct installers for certain versions (see 74ca299).
These installers should be correct for their given versions. I haven't tested
installing them, but timestamps with the digital signatures are about (one or
two days out sometimes) the release dates for the given versions, which would
strongly suggest that they are the correct installers.

Included installers are for the following version: v0.8, v0.9, v0.95.